A little performance improvment
The edited GetSeName method is called in other methods, some of them are during saving process of admin pages and some of them occur during user interactions like viewing tags. The ToCharArray method is called for every character of the name string, this update moves the ToCharArray method out of foreach loop.