Fixing the bug: "Notify for qty below" setting in product details doesn't work
The "Notify for qty below" setting in product details doesn't work. It doesn't notify the store owner for stock quantities below the specified value. This is because wrong values are used in queries instead of using "Notify for qty below" setting.
Queries get corrected to use the "Notify for qty below" setting (NotifyAdminForQuantityBelow) for filtering the result sets.