Do not check for store specifics to maintain original behavior
Originally, even if an Entity was not assigned to your current store, you could always Preview it in the current store.
This changes brings that back by displaying all the stores available.

This also gets around an issue where an Entity that was `LimitedTo` a deleted store had no preview options